+++ src/components/Home/Confirm/ConfirmWrite/ConfirmWriteImageSlideInput.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import 'swiper/css/pagination';
 import { Pagination } from 'swiper';
 import { Swiper, SwiperSlide } from 'swiper/react';
-import { ReactComponent as ConfirmWriteSliderSVG } from '../../../styles/images/icons/confirm-write__slider.svg';
+import { ReactComponent as ConfirmWriteSliderSVG } from '../../../../styles/images/icons/confirm-write__slider.svg';
 import { IConfirmWriteImageSlideInputProps, IImageFile } from 'types/Home';
 
 function ConfirmWriteImageSlideInput({
